requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
redis